Reorganize all files and seperated tests and methods
