--- lib/modules/team_stats.rb
@@ -1,3 +1,0 @@
-module TeamStats
-
-end
