--- test/game_teams_test.rb
@@ -1,1 +1,0 @@
-require './test/test_helper'
